--- a/library/lcd/lcd_comm_rev_b.py
+++ b/library/lcd/lcd_comm_rev_b.py
@@ -17,12 +17,10 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import struct
-
 from serial.tools.list_ports import comports
-import numpy as np
 
 from library.lcd.lcd_comm import *
+from library.lcd.serialize import image_to_RGB565, chunked
 from library.log import logger
 
 
@@ -195,38 +193,12 @@ class LcdCommRevB(LcdComm):
         else:
             self.SendCommand(Command.SET_ORIENTATION, payload=[OrientationValueRevB.ORIENTATION_LANDSCAPE])
 
-    @staticmethod
-    def imageToRGB565BE(image: Image.Image):
-        if image.mode not in ["RGB", "RGBA"]:
-            # we need the first 3 channels to be R, G and B
-            image = image.convert("RGB")
-
-        rgb = np.asarray(image)
-
-        # flatten the first 2 dimensions (width and height) into a single stream
-        # of RGB pixels
-        rgb = rgb.reshape((image.size[1] * image.size[0], -1))
-
-        # extract R, G, B channels and promote them to 16 bits
-        r = rgb[:, 0].astype(np.uint16)
-        g = rgb[:, 1].astype(np.uint16)
-        b = rgb[:, 2].astype(np.uint16)
-
-        # construct RGB565
-        r = (r >> 3)
-        g = (g >> 2)
-        b = (b >> 3)
-        rgb565 = (r << 11) | (g << 5) | b
-
-        # serialize to big-endian
-        return rgb565.astype('>u2').tobytes()
-
     def serialize_image(self, image: Image.Image, height: int, width: int) -> bytes:
         if image.width != width or image.height != height:
             image = image.crop((0, 0, width, height))
         if self.orientation == Orientation.REVERSE_PORTRAIT or self.orientation == Orientation.REVERSE_LANDSCAPE:
             image = image.rotate(180)
-        return self.imageToRGB565BE(image)
+        return image_to_RGB565(image, "big")
 
     def DisplayPILImage(
             self,
@@ -271,12 +243,5 @@ class LcdCommRevB(LcdComm):
         # Lock queue mutex then queue all the requests for the image data
         with self.update_queue_mutex:
             # Send image data by multiple of "display width" bytes
-            start = 0
-            end = self.get_width() * 8
-            while end <= len(rgb565be):
-                self.SendLine(rgb565be[start:end])
-                start, end = end, end + self.get_width() * 8
-
-            # Write last line if needed
-            if start != len(rgb565be):
-                self.SendLine(rgb565be[start:])
+            for chunk in chunked(rgb565be, self.get_width() * 8):
+                self.SendLine(chunk)