Extend the use of numpy for serializing the image data of revisions B, C and D. Cf. #319 

I've added tests to make sure the new serialization results in exactly the same calls to the serial interface as before. The tests can be run like this: `python -m unittest` from the root directory.

The tests can also be used as simple benchmarks (in that case the assertions are not checked) like this:

`BENCHMARK=1 python -m unittest`

On my machine, the benchmark results are as follows (before -> after)

* Rev B: 0.56 s -> 0.009 s **(~ 60x speedup)**
* Rev C: 1.45 s -> 0.036 s **(~ 40x speedup)**
* Rev D: 0.36 s -> 0.015 s **(~ 20x speedup)**

This PR is based off of #622 so ideally that one should be merged before this one.