@mathoudebine 2025-01-08
Thank you @hchargois much easier to read!
I will take some time to test it with the different displays before merging it