@mathoudebine 2025-02-12
@hchargois I successfully tested your PR against the 3 display revisions, everything works fine! 
It actually helped me find an issue with the rev. B displays failing to display bitmaps properly when frames are sent too fast 😄 I'm merging your PR and will add a fix for the rev. B afterwards
Thanks again for your work!